Change Pulldown so it doesn't need to be instantiated with new
No need for it to be a "class", so just made it a plain old boring
object. Much less overhead now, no need to call new Pulldown().init(),
just pulldown.init;

Renamed module from Pulldown -> pulldown.

Updated the CLI module and the tests to use the new way.

Deleted the files in lib, just put the source into the main file to keep
it tidy.